+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - feat: adds the `disableCertificateTimeVerification` optional field to the `CanisterStatus.request` function argument, which allows you to control the `disableTimeVerification` option for the internal `Certificate.create` call.
 - fix: avoid bigint overflow when decoding the time from the certificate.
 - fix: enable certificate freshness checks for delegation certificates.
+- feat: adds the `UncertifiedRejectUpdateErrorCode` error code.
+- fix: changes error code to `UncertifiedRejectUpdateErrorCode` when throwing an error in v2 responses.
+- fix: avoid syncing time indefinitely in case of an ingress expiry error.
+- fix: throw an error with code `UncertifiedRejectUpdateErrorCode` if the reply from the update call was returned undefined and the method has no return type.
 
 ## [3.1.0] - 2025-07-24